Add extra check for gbq job completion
Sometimes the gbq query_reply contains the 'jobComplete' key, but the value is False.  We only want to start parsing results when the query job is truly finished, otherwise we'll prematurely break out of the while loop and get a cryptic KeyError when the 'totalRows' key can't be found in query_reply.